+++ src/Smartstore.Core/Migrations/SmartDbContextDataSeeder.cs
@@ -57,7 +57,10 @@ namespace Smartstore.Core.Data.Migrations
 
         public void MigrateLocaleResources(LocaleResourcesBuilder builder)
         {
-            builder.Delete("Account.ChangePassword.Errors.PasswordIsNotProvided");
+            builder.Delete(
+                "Account.ChangePassword.Errors.PasswordIsNotProvided",
+                "Common.Wait...",
+                "Topic.Button");
 
             builder.AddOrUpdate("Admin.Report.MediaFilesSize", "Media size", "Mediengröße");
             builder.AddOrUpdate("Admin.Rules.FilterDescriptor.Affiliate", "Affiliate", "Partner");
@@ -79,6 +82,14 @@ namespace Smartstore.Core.Data.Migrations
             // Typo.
             builder.AddOrUpdate("Admin.Configuration.Settings.ShoppingCart.ShowConfirmOrderLegalHint.Hint")
                 .Value("de", "Legt fest, ob rechtliche Hinweise in der Warenkorbübersicht auf der Bestellabschlussseite angezeigt werden. Dieser Text kann in den Sprachresourcen geändert werden.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.UseNativeNameInLanguageSelector",
+                "Display native language name in language selector",
+                "In der Sprachauswahl den Sprachnamen in der Landesprache anzeigen",
+                "Specifies whether the native language name should be displayed in the language selector. Otherwise, the name maintained in the backend is used.",
+                "Legt fest, ob in der Sprachauswahl die Sprachnamen in der nativen Landesprache angezeigt werden soll. Ansonsten wird der im Backend hinterlegte Name verwendet.");
+
+            builder.AddOrUpdate("Common.PageNotFound", "The page does not exist.", "Die Seite existiert nicht.");
         }
     }
 }