+++ src/Smartstore.Web/Controllers/TopicController.cs
@@ -12,20 +12,22 @@ namespace Smartstore.Web.Controllers
 {
     public class TopicController : PublicController
     {
+        private readonly SmartDbContext _db;
         private readonly IAclService _aclService;
         private readonly IStoreMappingService _storeMappingService;
         private readonly SeoSettings _seoSettings;
         private readonly IBreadcrumb _breadcrumb;
         private readonly CatalogHelper _helper;
-        //private readonly ICookieManager _cookieManager;
 
         public TopicController(
+            SmartDbContext db,
             IAclService aclService,
             IStoreMappingService storeMappingService,
             SeoSettings seoSettings,
             IBreadcrumb breadcrumb,
             CatalogHelper helper)
         {
+            _db = db;
             _aclService = aclService;
             _storeMappingService = storeMappingService;
             _seoSettings = seoSettings;
@@ -33,9 +35,83 @@ namespace Smartstore.Web.Controllers
             _helper = helper;
         }
 
-        protected TopicModel PrepareTopicModel(Topic topic)
+        public async Task<IActionResult> TopicDetails(int topicId, bool popup = false)
         {
-            Guard.NotNull(topic, nameof(topic));
+            await _helper.GetBreadcrumbAsync(_breadcrumb, ControllerContext);
+
+            var cacheKey = string.Format(ModelCacheInvalidator.TOPIC_BY_ID_KEY,
+                topicId,
+                Services.WorkContext.WorkingLanguage.Id,
+                Services.StoreContext.CurrentStore.Id,
+                Services.WorkContext.CurrentCustomer.GetRolesIdent());
+
+            var cacheModel = await Services.CacheFactory.GetMemoryCache().GetAsync(cacheKey, async (o) =>
+            {
+                o.ExpiresIn(TimeSpan.FromDays(1));
+
+                var topic = await _db.Topics.FindByIdAsync(topicId, false);
+                if (topic != null
+                    && topic.IsPublished
+                    && await _storeMappingService.AuthorizeAsync(topic)
+                    && await _aclService.AuthorizeAsync(topic))
+                {
+                    return PrepareTopicModel(topic, !topic.IsPasswordProtected);
+                }
+
+                return null;
+            });
+
+            if (cacheModel == null || (!popup && cacheModel.RenderAsWidget))
+            {
+                return NotFound();
+            }
+
+            ViewBag.IsPopup = popup;
+
+            if (!cacheModel.RenderAsWidget)
+            {
+                Services.DisplayControl.Announce(new Topic { Id = cacheModel.Id });
+            }
+
+            return View(cacheModel);
+        }
+
+        [HttpPost, DisallowRobot]
+        public async Task<IActionResult> Authenticate(int id, string password)
+        {
+            var success = false;
+            var message = string.Empty;
+            var content = string.Empty;
+            var topic = await _db.Topics.FindByIdAsync(id, false);
+
+            if (topic != null 
+                && topic.IsPublished
+                && topic.IsPasswordProtected
+                && await _storeMappingService.AuthorizeAsync(topic)
+                && await _aclService.AuthorizeAsync(topic))
+            {
+                if (topic.Password != null && password != null && password == topic.Password)
+                {
+                    var model = PrepareTopicModel(topic, true);
+                    content = await InvokePartialViewAsync("_Topic", model);
+                    success = true;
+                }
+                else
+                {
+                    message = T("Topic.WrongPassword");
+                }
+            }
+            else
+            {
+                message = T("Common.PageNotFound");
+            }
+
+            return Json(new { success, message, content });
+        }
+
+        protected TopicModel PrepareTopicModel(Topic topic, bool includePasswordProtectedFields)
+        {
+            Guard.NotNull(topic);
 
             var titleTag = "h3";
             if (topic.TitleTag != null)
@@ -54,10 +130,10 @@ namespace Smartstore.Web.Controllers
                 HtmlId = topic.HtmlId,
                 BodyCssClass = topic.BodyCssClass,
                 IsPasswordProtected = topic.IsPasswordProtected,
-                ShortTitle = topic.IsPasswordProtected ? null : topic.GetLocalized(x => x.ShortTitle),
-                Title = topic.IsPasswordProtected ? null : topic.GetLocalized(x => x.Title),
-                Intro = topic.IsPasswordProtected ? null : topic.GetLocalized(x => x.Intro),
-                Body = topic.IsPasswordProtected ? null : topic.GetLocalized(x => x.Body, detectEmptyHtml: true),
+                ShortTitle = includePasswordProtectedFields ? topic.GetLocalized(x => x.ShortTitle) : null,
+                Title = includePasswordProtectedFields ? topic.GetLocalized(x => x.Title) : null,
+                Intro = includePasswordProtectedFields ? topic.GetLocalized(x => x.Intro) : null,
+                Body = includePasswordProtectedFields ? topic.GetLocalized(x => x.Body, detectEmptyHtml: true): null,
                 MetaKeywords = topic.GetLocalized(x => x.MetaKeywords),
                 MetaDescription = topic.GetLocalized(x => x.MetaDescription),
                 MetaTitle = topic.GetLocalized(x => x.MetaTitle),
@@ -72,78 +148,6 @@ namespace Smartstore.Web.Controllers
             }
 
             return model;
-        }
-
-        public async Task<IActionResult> TopicDetails(int topicId, bool popup = false)
-        {
-            await _helper.GetBreadcrumbAsync(_breadcrumb, ControllerContext);
-
-            var cacheKey = string.Format(ModelCacheInvalidator.TOPIC_BY_ID_KEY,
-                topicId,
-                Services.WorkContext.WorkingLanguage.Id,
-                Services.StoreContext.CurrentStore.Id,
-                Services.WorkContext.CurrentCustomer.GetRolesIdent());
-
-            var cacheModel = await Services.CacheFactory.GetMemoryCache().GetAsync(cacheKey, async (o) =>
-            {
-                o.ExpiresIn(TimeSpan.FromDays(1));
-
-                var topic = await Services.DbContext.Topics.FindByIdAsync(topicId, false);
-                if (topic == null || !topic.IsPublished)
-                    return null;
-
-                if (!await _storeMappingService.AuthorizeAsync(topic))
-                    return null;
-
-                if (!await _aclService.AuthorizeAsync(topic))
-                    return null;
-
-                return PrepareTopicModel(topic);
-            });
-
-            if (cacheModel == null || (!popup && cacheModel.RenderAsWidget))
-                return NotFound();
-
-            ViewBag.IsPopup = popup;
-
-            if (!cacheModel.RenderAsWidget)
-            {
-                Services.DisplayControl.Announce(new Topic { Id = cacheModel.Id });
-            }
-
-            return View("TopicDetails", cacheModel);
-        }
-
-        [HttpPost, DisallowRobot]
-        public async Task<IActionResult> Authenticate(int id, string password)
-        {
-            var authResult = false;
-            var title = string.Empty;
-            var body = string.Empty;
-            var error = string.Empty;
-            var db = Services.DbContext;
-
-            var topic = await db.Topics.FindByIdAsync(id, false);
-
-            if (topic != null &&
-                topic.IsPublished &&
-                topic.IsPasswordProtected &&
-                await _storeMappingService.AuthorizeAsync(topic) &&
-                await _aclService.AuthorizeAsync(topic))
-            {
-                if (topic.Password != null && topic.Password.Equals(password))
-                {
-                    authResult = true;
-                    title = topic.GetLocalized(x => x.Title).Value;
-                    body = topic.GetLocalized(x => x.Body, detectEmptyHtml: true).Value;
-                }
-                else
-                {
-                    error = T("Topic.WrongPassword").Value;
-                }
-            }
-
-            return Json(new { Authenticated = authResult, Title = title, Body = body, Error = error });
         }
     }
 }