+++ src/Smartstore/Imaging/ImagingHelper.cs
@@ -1,7 +1,6 @@
 ﻿using System.Drawing;
 using System.Globalization;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
+using System.Text.RegularExpressions;
 using SixLabors.ImageSharp.PixelFormats;
 using SharpColor = SixLabors.ImageSharp.Color;
 
@@ -9,10 +8,8 @@ namespace Smartstore.Imaging
 {
     public static class ImagingHelper
     {
-        public static ILogger Logger { get; set; } = NullLogger.Instance;
-
-        private static readonly string CssColorComponentDelimiterRegex = @"(\s*/\s*)|(\s*,\s*)|(\s+)";
-        private static readonly Color ColorTranslationFallback = Color.White;
+        private static readonly Regex ColorComponentDelimiterRegex = new(@"(\s*/\s*)|(\s*,\s*)|(\s+)", RegexOptions.Compiled);
+        private static readonly Color DefaultColor = Color.White;
 
         /// <summary>
         /// Converts a <see cref="SixLabors.ImageSharp.Color"/> to <see cref="System.Drawing.Color"/>.
@@ -41,22 +38,22 @@ namespace Smartstore.Imaging
         /// <returns>The <see cref="Color"/>.</returns>
         public static Color TranslateColor(string htmlColor)
         {
-            Guard.NotEmpty(htmlColor, nameof(htmlColor));
+            Guard.NotEmpty(htmlColor);
 
-            if (!SharpColor.TryParse(htmlColor, out var sharpColor))
+            if (!TryTranslateColor(htmlColor, out var color))
             {
                 throw new ArgumentException("Input string color is not in the correct format.", nameof(htmlColor));
             }
 
-            return ConvertColor(sharpColor);
+            return color;
         }
 
         /// <summary>
         /// Attempts to create a new instance of the <see cref="Color"/> struct from the given input string.
         /// </summary>
         /// <param name="htmlColor">
-        /// The name of the color or the hexadecimal representation of the combined color components arranged
-        /// in rgb, rgba, rrggbb, or rrggbbaa format to match web syntax.
+        /// The name of the color, or the hexadecimal representation of the combined color components arranged
+        /// in rgb, rgba, rrggbb, or rrggbbaa format to match web syntax, or the CSS rgb(a) representation.
         /// </param>
         /// <param name="result">When this method returns, contains the <see cref="Color"/> equivalent of the html input.</param>
         /// <returns>The <see cref="bool"/>.</returns>
@@ -75,49 +72,26 @@ namespace Smartstore.Imaging
                 return true;
             }
 
+            if ((htmlColor.StartsWith("rgb(") || htmlColor.StartsWith("rgba(")) && htmlColor.EndsWith(')'))
+            {
+                var colorFromRgbaString = ParseRgbaColor(htmlColor);
+                if (colorFromRgbaString == null)
+                {
+                    return false;
+                }
+
+                result = colorFromRgbaString.Value;
+                return true;
+            }
+
             return false;
         }
 
         public static int GetPerceivedBrightness(string htmlColor)
         {
-            Color colorFromHtml;
-
-            if (string.IsNullOrEmpty(htmlColor))
+            if (!TryTranslateColor(htmlColor, out var colorFromHtml))
             {
-                colorFromHtml = ColorTranslationFallback;
-            }
-            else
-            {
-                try
-                {
-                    colorFromHtml = ColorTranslator.FromHtml(htmlColor);
-                }
-                catch
-                {
-                    // Either an unknown color name, a CSS function, or an invalid hex string.
-                    colorFromHtml = Color.Empty;
-                }
-
-                if (colorFromHtml.IsEmpty)
-                {
-                    try
-                    {
-                        if ((htmlColor.StartsWithNoCase("rgb(") || htmlColor.StartsWithNoCase("rgba(")) && htmlColor.EndsWith(')'))
-                        {
-                            colorFromHtml = ConvertRgbaCssColor(htmlColor);
-                        }
-                        else
-                        {
-                            // Invalid or unknown color name / function. Use fallback color.
-                            colorFromHtml = ColorTranslationFallback;
-                        }
-                    }
-                    catch
-                    {
-                        // Invalid or too complex color code. Use fallback color.
-                        colorFromHtml = ColorTranslationFallback;
-                    }
-                }
+                colorFromHtml = DefaultColor;
             }
 
             return GetPerceivedBrightness(colorFromHtml);
@@ -127,25 +101,25 @@ namespace Smartstore.Imaging
         /// Converts a CSS rgba() color string into a <see cref="Color"/> instance.
         /// </summary>
         /// <param name="htmlColor">The CSS color string must be in rgb(r g b) or rgba(r g b a) format. Valid delimiters are space, comma, and slash.</param>
-        /// <returns>The <see cref="Color"/>.</returns>
-        private static Color ConvertRgbaCssColor(string htmlColor)
+        /// <returns>The <see cref="Color"/> or <c>null</c> if input is malformed.</returns>
+        private static Color? ParseRgbaColor(string htmlColor)
         {
             var rgba = htmlColor.Substring(htmlColor.IndexOf('(') + 1, htmlColor.IndexOf(')') - htmlColor.IndexOf('(') - 1);
 
-            // Separate the values by spaces and /or commas.
-            rgba = rgba.RegexReplace(CssColorComponentDelimiterRegex, " ");
+            // Separate the values by spaces and/or commas.
+            rgba = ColorComponentDelimiterRegex.Replace(rgba, " ");
             var rgbParts = rgba.Split(' ');
 
             // Convert the values to integers.
-            var r = ConvertCssColorComponent(rgbParts[0]);
-            var g = ConvertCssColorComponent(rgbParts[1]);
-            var b = ConvertCssColorComponent(rgbParts[2]);
-            var a = rgbParts.Length == 3 ? 255 : ConvertCssColorComponent(rgbParts[3]);
+            var r = ConvertColorComponent(rgbParts[0]);
+            var g = ConvertColorComponent(rgbParts[1]);
+            var b = ConvertColorComponent(rgbParts[2]);
+            var a = rgbParts.Length == 3 ? 255 : ConvertColorComponent(rgbParts[3]);
 
             // On error, use fallback color.
             if (r == null || g == null || b == null || a == null)
             {
-                return ColorTranslationFallback;
+                return null;
             }
             else
             {
@@ -158,12 +132,12 @@ namespace Smartstore.Imaging
         /// </summary>
         /// <param name="colorComponent">The CSS color component can be an integer, a double, or a percentage.</param>
         /// <returns>An <see cref="int"/> between 0 and 255. Invalid values default to 0.</returns>
-        private static int? ConvertCssColorComponent(string colorComponent)
+        private static int? ConvertColorComponent(string colorComponent)
         {
             // Check for percentage values.
             if (colorComponent.EndsWith('%'))
             {
-                if (!double.TryParse(colorComponent.Substring(0, colorComponent.Length - 1), NumberStyles.Any, CultureInfo.InvariantCulture, out double doubleVal))
+                if (!double.TryParse(colorComponent[..^1], NumberStyles.Any, CultureInfo.InvariantCulture, out double doubleVal))
                 {
                     return null;
                 }
@@ -221,7 +195,7 @@ namespace Smartstore.Imaging
         /// <returns>The rescaled size</returns>
         public static Size Rescale(Size original, int maxSize)
         {
-            Guard.IsPositive(maxSize, nameof(maxSize));
+            Guard.IsPositive(maxSize);
 
             return Rescale(original, new Size(maxSize, maxSize));
         }